requirements.txt
django==4.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
python-slugify==8.0.4
python-dotenv==1.0.1
django-phonenumber-field==8.0.0
wonderwords==2.2.0
phonenumbers==8.13.54
django-debug-toolbar==4.3.0